[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -3,13 +3,13 @@
 using namespace std;
 int main()
 {
-    int a;
-    int b;
+    long long a;
+    long long b;
     
     cin>>a;
     cin>>b;
     
-    int result;
+    long result;
     result = (a+b)*(a-b);
     
     cout<<result;
